requirements.txt
pip==24.0
paho-mqtt==1.6.1